athena/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.3.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
httpx==0.27.0
apscheduler==3.10.4
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
ccxt==4.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.0
